Fix business hours day name mismatch
- Change from weekday: 'short' (Mon, Tue) to weekday: 'long' (monday, tuesday)
- Database stores full day names: monday, tuesday, wednesday, etc.
- Component was looking for short names - causing businessHours to not render

This fixes the issue where business hours section didn't display on landing page.

+++ components/subdomain/BusinessHoursDisplay.tsx
@@ -35,7 +35,7 @@ export default function BusinessHoursDisplay({
     );
   }
 
-  const currentDay = new Date().toLocaleDateString('en-US', { weekday: 'short' }).toLowerCase();
+  const currentDay = new Date().toLocaleDateString('en-US', { weekday: 'long' }).toLowerCase();
   const todayHours = businessHours[currentDay];
   const isOpenToday = todayHours?.isOpen || false;
